requirements.txt
easydict

numpy
sklearn
matplotlib
cv2
open3d
plyfile

torch_geometric